+++ doc/whats_new.rst
@@ -295,6 +295,13 @@ API changes summary
      Previously it was broken for input of non-int dtype and the weighted
      array that was returned was wrong. By `Manoj Kumar`_.
 
+   - Change default value of precompute in :class:`ElasticNet`
+     :class:`ElasticNetCV`, :class:`Lasso` and :class:`LassoCV` from "auto"
+     to False. Setting precompute to "auto" was found to be slower since
+     the computation of the Gram matrix is computationally expensive and
+     outweighs the benefit of fitting the Gram.
+     By `Manoj Kumar`_.
+
 .. _changes_0_14:
 
 0.14